requirements.txt
flask
flask-socketio 
eventlet
gevent
gevent-websocket
Flask-SQLAlchemy
langchain
sentence_transformers
python-dotenv